+++ app/src/main/java/com/bardur/moviedb/api/MovieDatabaseApiService.kt
@@ -13,7 +13,7 @@ import retrofit2.http.Query
 private const val BASE_URL = "https://api.themoviedb.org/3/"
 
 // get your API Key from https://www.themoviedb.org/ replace [MY_API_KEY]
-private const val API_KEY = ""
+private const val API_KEY = "43ad9df8f9e03181f4ca41b5e00c8222"
 
 /**
  * Build the Moshi object that Retrofit will be using, making sure to add the Kotlin adapter for
@@ -33,12 +33,7 @@ private val retrofit = Retrofit.Builder()
     .build()
 
 interface MovieDatabaseApiService {
-    /**
-     * Returns a Coroutine [List] of [Movie] which can be fetched with await() if
-     * in a Coroutine scope.
-     * The @GET annotation indicates that the "mostPopular" endpoint will be requested with the GET
-     * HTTP method
-     */
+
     @GET("discover/movie?api_key=$API_KEY&sort_by=popularity.desc")
     suspend fun mostPopular(): MovieDatabaseResponseList
 
@@ -50,8 +45,6 @@ interface MovieDatabaseApiService {
 
     @GET("search/movie?api_key=$API_KEY")
     suspend fun search(@Query("query") query: String): MovieDatabaseResponseList
-
-
 
 }
