+++ app/src/main/java/com/bardur/moviedb/ui/screens/details/DetailsFragment.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -12,13 +11,14 @@ import com.bardur.moviedb.MainActivity
 import com.bardur.moviedb.R
 import com.bardur.moviedb.databinding.FragmentDetailsBinding
 import com.bardur.moviedb.storage.MovieStorageRepo
+import com.bardur.moviedb.ui.utills.ApplicationViewModelFactory
 
 class DetailsFragment : Fragment() {
 
     private lateinit var binding: FragmentDetailsBinding
 
     private lateinit var detailsViewModel: DetailsViewModel
-    private lateinit var viewModelFactory: DetailsViewModelFactory
+    private lateinit var viewModelFactory: ApplicationViewModelFactory
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -31,7 +31,10 @@ class DetailsFragment : Fragment() {
         binding = DataBindingUtil.inflate(inflater, R.layout.fragment_details, container, false)
 
         viewModelFactory =
-            DetailsViewModelFactory(DetailsFragmentArgs.fromBundle(requireArguments()).movie, repo)
+            ApplicationViewModelFactory(
+                movie = DetailsFragmentArgs.fromBundle(requireArguments()).movie,
+                movieStorageRepo = repo
+            )
 
         detailsViewModel =
             ViewModelProvider(this, viewModelFactory).get(DetailsViewModel::class.java)
